CoreConfigManager class
