+++ Demo/Demo/DemoSettings/DemoSettings.swift
@@ -43,20 +43,4 @@ enum DemoSettings {
             UserDefaults.standard.set(newValue.rawValue, forKey: MerchantIntegrationDefaultKey)
         }
     }
-
-    static var clientID: String {
-        switch (environment, merchantIntegration) {
-        case (.sandbox, .unspecified):
-            return "AcXwOk3dof7NCNcriyS8RVh5q39ozvdWUF9oHPrWqfyrDS4AwVdKe32Axuk2ADo6rI_31Vv6MGgOyzRt"
-        case (.sandbox, .direct):
-            return "AVhcAP8TDu5PFeAw97M8187g-iYQW8W0AhvvXaMaWPojJRGGkunX8r-fyPkKGCv09P83KC2dijKLKwyz"
-        case (.sandbox, .connectedPath):
-            return "AcvkeOozOElJtQoZxxdAsDUrsClbAiNv7KIW6675dAiC7EX5R0wvSPiUNCc2JPEKHyFPfegwh_OV2afV"
-        case (.sandbox, .managedPath):
-            return "Afba8WbtSOqWlMoxroE5Ym8CVnogJJcHpj2uFpPYzN7oJz8NOi9XRrmHmpbWVQm6vRX0SwCvabaNKo06"
-        case (.live, _):
-            // TODO: Investigate getting a prod testing account that doesn't charge real money
-            return "TODO"
-        }
-    }
 }
